edge-ai-backend-main/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
pandas
torch
onnxruntime
Pillow

--- 🔽 AI 기능 라이브러리 추가 ---
opencv-python==4.8.0.76
mediapipe==0.10.20
ultralytics